+++ app/src/main/java/com/e/amicummobile/application/TranslatorApp.kt
@@ -4,6 +4,7 @@ import android.app.Application
 import com.e.amicummobile.di.application
 import com.e.amicummobile.di.db
 import com.e.amicummobile.di.mainScreen
+import com.e.amicummobile.di.notification
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.GlobalContext.startKoin
 
@@ -16,7 +17,8 @@ class TranslatorApp : Application() {
                 listOf(
                     application,
                     db,
-                    mainScreen
+                    mainScreen,
+                    notification
                 )
             )
         }
