+++ app/src/main/java/com/e/amicummobile/di/KoinModules.kt
@@ -3,6 +3,7 @@ package com.e.amicummobile.di
 import androidx.room.Room
 import com.example.utils.network.Network
 import com.e.amicummobile.interactor.MainInteractor
+import com.e.amicummobile.view.notification.StoreNotification
 import com.e.amicummobile.viewmodel.StoreAmicum
 import com.example.db.AmicumDB
 import com.example.repository.IRepositoryLocal
@@ -10,6 +11,7 @@ import com.example.repository.IRepositoryRemote
 import com.example.repository.localRepository.RoomRepository
 import com.example.repository.localRepository.TestDataRepository
 import org.koin.android.ext.koin.androidContext
+import org.koin.core.qualifier.named
 import org.koin.dsl.module
 
 /**
@@ -31,5 +33,10 @@ val mainScreen = module {
 val db = module {
     single { Room.databaseBuilder(get(), AmicumDB::class.java, "AmicumDB").build() }          // база данных
     single { get<AmicumDB>().handbooksDao() }                                                       // доступ к справочникам системы
+}
 
+val notification = module {
+    scope(named("NOTIFICATION_STORE")) {
+        scoped { StoreNotification(get(), get()) }
+    }
 }
